requirements.txt
networkx==3.1
pyvis==0.3.1
coverage==7.6.1
matplotlib==3.7.5
pyyaml==6.0